Account for different vector lengths when interpolating

+++ src/util/VectorUtil.h
@@ -1,3 +1,3 @@
 #include "definitions/FaceVectors.h"
 
-static void interpolateVector(vector<Point>& currentVector, vector<Point>& targetVector, float amount);
+static void interpolateVector(vector<Point>& currentVector, vector<Point>& targetVector, double amount);
